Clean up imports and exports to help with package and bundle size (#721)
