+++ ui/plugin-system/package.json
@@ -14,6 +14,7 @@
   "module": "dist/index.js",
   "main": "dist//cjs/index.js",
   "types": "dist/index.d.ts",
+  "sideEffects": false,
   "scripts": {
     "clean": "rimraf dist/",
     "build": "concurrently \"npm:build:*\"",
